cleanup, renaming types, removing redundant code
