+++ modules/server/src/network/health.ts
@@ -1,6 +1,5 @@
 import axios from 'axios';
 import urljoin from 'url-join';
-import { CONNECTION_STATUS } from './resolvers/nodes';
 import { ConnectionStatus } from './types';
 
 /**
