+++ modules/server/src/network/types/setup.ts
@@ -1,4 +1,7 @@
-// environment config
+/*
+ * Setup of network types
+ */
+
 export type NetworkConfig = {
 	graphqlUrl: string;
 	documentType: string;
