+++ modules/server/src/network/types/types.ts
@@ -1,6 +1,7 @@
 import { ObjectValues } from '@/utils/types';
-import { SupportedAggregation } from './common';
-import { CONNECTION_STATUS } from './resolvers/remoteConnections';
+import { SupportedAggregation } from '../common';
+import { CONNECTION_STATUS } from '../resolvers/networkNode';
+import { Aggregations, Bucket, NumericAggregations } from './aggregations';
 
 // environment config
 export type NetworkAggregationConfigInput = {
@@ -24,29 +25,6 @@ export type SupportedAggregations = SupportedNetworkFieldType[];
 export type UnsupportedAggregations = NetworkFieldType<string>[];
 
 export type ConnectionStatus = ObjectValues<typeof CONNECTION_STATUS>;
-
-export type RemoteConnectionData = {
-	url: string;
-	name: string;
-	description: string;
-	documentName: string;
-	availableAggregations: SupportedNetworkFieldType[];
-	totalHits: number;
-	errors: string[];
-	status: ConnectionStatus;
-};
-
-export type Bucket = {
-	doc_count: number;
-	key: string;
-};
-
-export type Aggregations = {
-	bucket_count: number;
-	buckets: Bucket[];
-};
-
-export type NumericAggregations = {};
 
 export type RemoteAggregation = { [key: string]: Aggregations | NumericAggregations };
